requirements.txt
ncatbot==3.5.9
Pillow==11.1.0
pyzbar==0.1.9
Requests==2.32.3
selenium==4.24.0
pyyaml~=6.0.2